% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMovie.R
\name{plotMovie}
\alias{plotMovie}
\title{Plot Movie based on Year}
\usage{
plotMovie(queryTitle, color = "green")
}
\arguments{
\item{queryTitle}{title of the movie}

\item{color}{color of specified movie in plot}
}
\value{
plot
}
\description{
Plot Movie based on Year
}
